Sat, Dec 30, 2023, 6:51 PM +08:00

+++ CS50P/interpreter/interpreter.py
@@ -1,10 +1,10 @@
 x, y, z= input("Expression: ").split(" ")
 
 if y == "+":
-    print(int(x) + int(z))
+    print(float(int(x) + int(z)))
 elif y == "-":
-    print(int(x) - int(z))
+    print(float(int(x) - int(z)))
 elif y == "*":
-    print(int(x) * int(z))
+    print(float(int(x) * int(z)))
 elif y == "/":
-    print(int(x) / int(z))
+    print(float(int(x) / int(z)))
